+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/Images/decodeCompressedImageToBitmap.ts
@@ -1,0 +1,13 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import { CompressedImageTypes } from "./ImageTypes";
+
+export async function decodeCompressedImageToBitmap(
+  image: CompressedImageTypes,
+  resizeWidth?: number,
+): Promise<ImageBitmap> {
+  const bitmapData = new Blob([image.data], { type: `image/${image.format}` });
+  return await createImageBitmap(bitmapData, { resizeWidth });
+}
